Move create custom page form to index
