+++ front/app/services/customPages.ts
@@ -1,0 +1,14 @@
+import { API_PATH } from 'containers/App/constants';
+import streams from 'utils/streams';
+import { Multiloc } from 'typings';
+
+interface ICustomPage {}
+
+const customPagesEndpoint = `${API_PATH}/pages`;
+
+export function createCustomPageStream(pageData: {
+  title_multiloc: Multiloc;
+  slug?: string;
+}) {
+  return streams.add<ICustomPage>(`${customPagesEndpoint}`, pageData);
+}
